requirements.txt
pymongo==4.11.2
python-dotenv==1.0.1
matplotlib==3.10.1
numpy==2.2.2
Pillow==11.1.0
